% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_combinations.R
\name{combinations}
\alias{combinations}
\alias{unique_combos}
\alias{get_combos}
\title{Get all patterns of diseases.}
\usage{
unique_combos(positions, combinations = 2)

get_combos(positions, combinations = 2)
}
\arguments{
\item{combinations}{Maximum number of disease combinations to be generated (default = 2)}

\item{positons}{Vector of all positions for each element of comorbid_column generated using .get_locales}
}
\value{
matrix of unique combinations (if unique_combos, else all combinations)
}
\description{
Return combinations of different patterns of diseases to a specified number of combinations
}
\examples{
positions = list(c(1,2,3,4), c(3), c(3), c(2,3))
unique_combos(positions, combinations = 2)

}
