% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbid_strings.R
\name{comorbid}
\alias{comorbid}
\alias{to_string}
\alias{check_strings_equal}
\alias{get_disease_counts}
\title{Helper functions to check comorbid_strings and get count position}
\usage{
to_string(comorbid_column)

check_strings_equal(comorbid_column)

get_disease_counts(comorbid_column)
}
\arguments{
\item{comorbid_column}{A vector of character strings made up of 0s and 1s, or of factors coercible to character, all should be identical lengths.}
}
\description{
Convert and check comorbid strings, count positions within them to get disease counts.
}
\examples{

comorbid_column <- c('00010', '01000', '01110', '11101')

comorbid_factors <- as.factor(comorbid_column)

to_string(comorbid_factors) # make into string

check_strings_equal(comorbid_column) # TRUE if strings all same length

get_disease_counts(comorbid_column) # list of diseases counts

}
