% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_names.R
\docType{data}
\name{disease_names}
\alias{disease_names}
\title{Names of diseases for hip_data}
\format{
A vector of 12 diseases
}
\source{
\url{https://digital.nhs.uk/data-and-information/publications/statistical/patient-reported-outcome-measures-proms/}
}
\usage{
disease_names
}
\description{
A vector containing the names of the 12 diseases captured in the hip_data.
}
\keyword{datasets}
