% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hip_data.R
\docType{data}
\name{hip_data}
\alias{hip_data}
\title{Example comorbidity and outcome data}
\format{
A data frame with two columns:
\describe{
  \item{comorbid_string}{A string of 0s and 1s indicating disease presence or absence, each position refers to a single disease}
  \item{outcome}{A binary 1/0 indicating if an outcome was suffered}
 }
}
\source{
\url{https://digital.nhs.uk/data-and-information/publications/statistical/patient-reported-outcome-measures-proms/}
}
\usage{
hip_data
}
\description{
A dataset containing the comorbid string of 40,000 individuals
and which suffered any complications after hip replacement surgery.
}
\keyword{datasets}
