% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_stems.R
\name{name_stems}
\alias{name_stems}
\title{Change split stems into named diseases}
\usage{
name_stems(split_stems, dis_names)
}
\arguments{
\item{split_stems}{A vector of split stems outputted by \code{\link{split_stem}}}

\item{dis_names}{A list of disease names associated with each position as used in \code{\link{get_disease_counts}}}
}
\value{
A character vector of split stems converted to names, with each disease in a stem separated by |.
}
\description{
Name stems once they have been split
}
\examples{
cut_stems = c('1-3', '2-3','1-2')

disease_names = list('Diabetes','Kidney Disease','Heart Failure')

name_stems(cut_stems, dis_names = disease_names) # many stems

}
