% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_funnel_plot.R
\name{stem_funnel_plot}
\alias{stem_funnel_plot}
\title{Funnel plot of combinations at a given stem level.}
\usage{
stem_funnel_plot(
  comorbid_column,
  outcome_column,
  cut_level = 2,
  use_outcome = FALSE,
  maximal = TRUE,
  max_x = NULL,
  dis_names = NULL
)
}
\arguments{
\item{comorbid_column}{A vector of character strings made up of 0s and 1s, or of factors coercible to character, all should be identical lengths.}

\item{outcome_column}{A numeric vector one if outcome occurred and zero if outcome did not occur. Should be the same length as \code{comorbid_column} with each element relating to the same record as the \code{comorbid_column}. If \code{outcome_column} is passed, then the stem will be generated based on combinations with the highest event rate.}

\item{cut_level}{A number indicating the level at which stems should be cut (i.e. the number of diagnoses).}

\item{use_outcome}{A logical indicating if stem derivation should be based on outcome rates or on frequency, defaults to \code{FALSE}.}

\item{maximal}{A logical that true if the 'maximal' stem is to be considered for each record, or if it is to be limited to only records with the same number of diagnoses as \code{cut_level}, default of \code{TRUE}.}

\item{max_x}{A number indicating the maximum frequency on the x axis, if left \code{NULL} then the maximum in the observed data will be used.}

\item{dis_names}{A list of disease names associated with each position of the comorbid string. If null, then the simple string is returned.}
}
\value{
A list, first element is a ggplot object; a funnel plot with associated 95\% and 99\% control limits, second element is the underlying data.
}
\description{
Create a funnel plot for a binary outcome for each combination of diseases. Makes heavy use of the \code{\link[funnelR:fundata]{funnelR:fundata}} and \code{\link[funnelR:funscore]{funnelR:funscore}} functionality from the \href{https://cran.r-project.org/web/packages/funnelR/funnelR.pdf}{funnelR} package.
}
\examples{

comorbid_column <- c('1110', '0010', '0010', '0110')

outcome_column <- c(1,0,0,1)

stem_funnel_plot(comorbid_column = comorbid_column,
outcome_column = outcome_column,
use_outcome = FALSE,
maximal = TRUE,
max_x = NULL,
dis_names = c('heart disease', 'hypertension', 'chronic kidney disease', 'cancer')
)

}
