% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_method.R
\name{summary.stem}
\alias{summary.stem}
\title{Summary method for stem}
\usage{
\method{summary}{stem}(stem, ...)
}
\arguments{
\item{object}{An object created by make_stem}
}
\description{
Summary method for stem
}
