% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_combinations.R
\name{combinations}
\alias{combinations}
\alias{get_combos}
\alias{unique_combos}
\title{Get all combinations and unique combinations}
\usage{
get_combos(positions, combinations = 2)

unique_combos(positions, combo_numbers = 2)
}
\arguments{
\item{combinations}{= 2}
}
\value{

}
\description{
thing to do
}
\examples{

}
