% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_default_combinations.R
\name{determine_default_combinations}
\alias{determine_default_combinations}
\title{Determine default number of combinations}
\usage{
determine_default_combinations(comorbid_column)
}
\value{

}
\description{
thing to determine best number of combinations
}
