% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stem.R
\name{make_stem}
\alias{make_stem}
\title{Make combinations of diseases ('stems') up to a maximum level.}
\usage{
make_stem(comorbid_column, max = "default")
}
\arguments{
\item{comorbid_column}{A vector of character strings made up of 0s and 1s, or of factors coercible to character, all should be identical lengths.}

\item{max}{The maximum number of combinations to be considered, if left as 'default' then a maximum based on code frequency will be used.}
}
\value{
data.frame with four columns, one row per unique comorbid_string pattern, comorbid pattern, Frequency of pattern, positions of diseases, and full stem.
}
\description{
Generate disease stems from a vector of character strings.
}
\examples{
comorbid_column <- c('00010', '01000', '01110', '11101')
make_stem(comorbid_column, max='default')

}
