% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions_plot.R
\name{stem_funnel_plot}
\alias{stem_funnel_plot}
\title{Funnel plot of proportions}
\usage{
stem_funnel_plot(comorbid_column, outcome, stem_cut = 2, x_lim = NULL)
}
\arguments{
\item{params}{add in}
}
\description{
description add in
}
