% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_generator.R
\name{stem_generator}
\alias{stem_generator}
\title{Generate disease stems}
\usage{
stem_generator(poscolumn, max_combos = 3, all_diseases, min_freq = 0)
}
\arguments{
\item{poscolumn}{Vector of numbers which describe position of 1s in a comorbid string, generated using get_locales().}

\item{max_combos}{Number of maximum combinations to attempt.}

\item{all_diseases}{List of positions associated with each disease, generated using get_disease_counts().}
}
\value{
main_stem for each pattern up to a maximum set in max_combos.
}
\description{
add something better here
}
